apps/agent/requirements.txt
livekit-agents>=1.0.0
livekit-plugins-silero>=1.0.0
livekit-plugins-deepgram>=1.0.0
livekit-plugins-cartesia>=1.0.0
livekit-plugins-openai>=1.0.0
anthropic>=0.42.0
pinecone-client>=3.0.0
numpy>=1.26.0
pydantic>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
